Add files via upload
Borja discovered that at previous commits, the backend directory was not uploaded because there were too many files. That amount of files were caused by the JavaScript libraries at the internal directory node_modules. 
In order to allow the upload of the backend code, that node_modules directory has not been uploaded. That means that at the pull of the code of github, the downloader shall have that code previously installed on its working directory.
At this upload, node_modules was installed for the EXPRESS and SQLITE3 libraries.